# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CrossWithFriends API",
    "description": "API for the CrossWithFriends crossword puzzle platform",
    "version": "1.0.0",
    "contact": {"name": "CrossWithFriends"}
  },
...
  "paths": {
    "/auth/token": {
      "post": {
        "operationId": "createToken",
        "summary": "Create authentication token",
        "tags": ["Authentication"],
        "description": "Creates a new JWT authentication token for a user. If userId is not provided, a new secure user ID is generated.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "Optional user ID. If not provided, a new one is generated."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["token", "userId", "expiresAt"],
                  "properties": {
                    "token": {"type": "string", "description": "JWT authentication token"},
                    "userId": {"type": "string", "description": "The user ID associated with this token"},
                    "expiresAt": {"type": "number", "description": "Token expiration timestamp (ms)"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "message"],
                  "properties": {
                    "error": {"type": "string", "description": "Error type"},
                    "message": {"type": "string", "description": "Error message"},
                    "statusCode": {"type": "number", "description": "HTTP status code"},
                    "retryAfter": {"type": "number", "description": "Retry after seconds (for rate limiting)"}
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "message"],
                  "properties": {
                    "error": {"type": "string", "description": "Error type"},
                    "message": {"type": "string", "description": "Error message"},
                    "statusCode": {"type": "number", "description": "HTTP status code"},
                    "retryAfter": {"type": "number", "description": "Retry after seconds (for rate limiting)"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/firebase-token": {
      "post": {
        "operationId": "exchangeFirebaseToken",
        "summary": "Exchange Firebase ID token for backend JWT",
        "tags": ["Authentication"],
        "description": "Verifies a Firebase ID token and returns a backend JWT token. The Firebase UID is used as the userId in the backend JWT.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["firebaseToken"],
                "properties": {
                  "firebaseToken": {"type": "string", "description": "Firebase ID token from the frontend"}
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["token", "userId", "expiresAt"],
                  "properties": {
                    "token": {"type": "string", "description": "Backend JWT authentication token"},
                    "userId": {"type": "string", "description": "The Firebase UID used as userId"},
                    "expiresAt": {"type": "number", "description": "Token expiration timestamp (ms)"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "message"],
                  "properties": {
                    "error": {"type": "string", "description": "Error type"},
                    "message": {"type": "string", "description": "Error message"},
                    "statusCode": {"type": "number", "description": "HTTP status code"},
                    "retryAfter": {"type": "number", "description": "Retry after seconds (for rate limiting)"}
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "message"],
                  "properties": {
                    "error": {"type": "string", "description": "Error type"},
                    "message": {"type": "string", "description": "Error message"},
                    "statusCode": {"type": "number", "description": "HTTP status code"},
                    "retryAfter": {"type": "number", "description": "Retry after seconds (for rate limiting)"}
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "message"],
                  "properties": {
                    "error": {"type": "string", "description": "Error type"},
                    "message": {"type": "string", "description": "Error message"},
                    "statusCode": {"type": "number", "description": "HTTP status code"},
                    "retryAfter": {"type": "number", "description": "Retry after seconds (for rate limiting)"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/validate": {
      "post": {
        "operationId": "validateToken",
        "summary": "Validate authentication token",
        "tags": ["Authentication"],
        "description": "Validates a JWT token and returns its payload if valid.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {"token": {"type": "string", "description": "The token to validate"}}
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["valid"],
                  "properties": {
                    "valid": {"type": "boolean", "description": "Whether the token is valid"},
                    "userId": {"type": "string", "description": "User ID from token (if valid)"},
                    "expiresAt": {"type": "number", "description": "Token expiration (if valid)"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "message"],
                  "properties": {
                    "error": {"type": "string", "description": "Error type"},
                    "message": {"type": "string", "description": "Error message"},
                    "statusCode": {"type": "number", "description": "HTTP status code"},
                    "retryAfter": {"type": "number", "description": "Retry after seconds (for rate limiting)"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "summary": "Get current user",
        "tags": ["Authentication"],
        "description": "Returns information about the currently authenticated user based on the Authorization header.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["userId", "expiresAt"],
                  "properties": {
                    "userId": {"type": "string", "description": "Current user ID"},
                    "expiresAt": {"type": "number", "description": "Token expiration timestamp"}
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "message"],
                  "properties": {
                    "error": {"type": "string", "description": "Error type"},
                    "message": {"type": "string", "description": "Error message"},
                    "statusCode": {"type": "number", "description": "HTTP status code"},
                    "retryAfter": {"type": "number", "description": "Retry after seconds (for rate limiting)"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/oembed": {
      "get": {
        "operationId": "getOembed",
        "summary": "Get oEmbed data",
        "tags": ["Link Preview"],
        "description": "Returns oEmbed format metadata for link previews",
        "parameters": [
          {
            "schema": {"type": "string"},
            "in": "query",
            "name": "author",
            "required": true,
            "description": "Author name"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "version", "author_name"],
                  "properties": {
                    "type": {"type": "string"},
                    "version": {"type": "string"},
                    "author_name": {"type": "string", "description": "Author name"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {"url": "https://www.crosswithfriends.com/api", "description": "Production server"},
    {"url": "https://crosswithfriendsbackend-staging.onrender.com/api", "description": "Staging server"},
    {"url": "http://localhost:3021/api", "description": "Local development server"}
  ],
...
}